BUG: impossible to import a med file of size more than MAX_INT (~2.1 Gb)   (SMESH_File.*)
Fix an error of porting to OCCT-7
  (StdMeshers_Cartesian_3D.cxx)

Remove useless MESSAGE's

+++ src/SMESHUtils/SMESH_File.hxx
@@ -109,7 +109,7 @@ public:
 private:
 
   std::string _name; //!< file name
-  int         _size; //!< file size
+  long        _size; //!< file size
   std::string _error;
 #ifdef WIN32
   HANDLE      _file, _mapObj;
